requirements.txt
django-background-tasks~=1.2.5
django~=3.2.5
pandas~=1.3.0
psycopg2~=2.9.1
pyyaml~=5.4.1